+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Added:** `itron::abi::acre_mpf` when `cfg(all(feature = "solid_asp3", feature = "dcre"))`
 - **Added:** `itron::abi::exd_tsk`
 - **Added:** `itron::task::Builder::finish_and_activate`
+- **Added:** `itron::task::TaskRef::migrate`
 - **Added:** `itron::memorypool::MemoryPool[Ref]`
 - **Added:** `itron::messagebuffer::MessageBuffer[Ref]`